+++ lib/figjam/rails/application.rb
@@ -1,19 +1,17 @@
 module Figjam
   module Rails
     class Application < Figjam::Application
-      private
-
-      def default_path
+      private def default_path
         rails_not_initialized unless ::Rails.root
 
         ::Rails.root.join("config", "application.yml")
       end
 
-      def default_environment
+      private def default_environment
         ::Rails.env
       end
 
-      def rails_not_initialized
+      private def rails_not_initialized
         raise RailsNotInitialized
       end
     end